\name{cepp-package}
\alias{cepp-package}
\alias{cepp}
\docType{package}

\title{Context Driven Exploratory Projection Pursuit.}

\description{Functions and Data for Context-driven Exploratory Projection Pursuit.}

\details{
\tabular{ll}{
Package: \tab cepp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-11\cr
License: \tab GPL - 3 \cr
}
}
\author{
Mohit Dayal

Maintainer: Mohit Dayal <mohitdayal2000@gmail.com>
}

\keyword{ projection pursuit }
