\name{cepp-package}
\alias{cepp-package}
\alias{cepp}
\docType{package}
\title{
\packageTitle{cepp}
}
\description{
\packageDescription{cepp}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{cepp}
}

\author{
\packageAuthor{cepp}

Maintainer: \packageMaintainer{cepp}
}

\keyword{projection pursuit}
